Cambio al nombre de la carpeta además creación de documentación y requisitos de la entrega
